documentation and some refactor
