+++ src/Components/Layout/layout.jsx
@@ -1,5 +1,11 @@
 import PropTypes from "prop-types";
 
+/**
+ * Renders a layout component with a scrollable container and centered content.
+ * @param {Object} props - The component props.
+ * @param {React.ReactNode} props.children - The child elements to be rendered inside the layout.
+ * @returns {JSX.Element} - The JSX element representing the layout component.
+ */
 const Layout = ({ children }) => {
   Layout.propTypes = {
     children: PropTypes.node.isRequired,
