+++ src/Contexts/hackerNewsContext.tsx
@@ -6,12 +6,25 @@ import { PostResultsSearch } from "../Models/post-results.model";
 const baseUrl = process.env.REACT_APP_BASE_URL;
 const searchtype: string = "search_by_date";
 
+/**
+ * Context object for the Hacker News app.
+ * @typedef {Object} HackerNewsContext
+ * @property {boolean} loadingApi - Indicates if the API is currently loading.
+ * @property {number} page - The current page number.
+ * @property {function} setPage - Function to set the current page number.
+ * @property {string} technologyType - The current technology type.
+ * @property {function} setTechnologyType - Function to set the current technology type.
+ * @property {Post[]} posts - Array of posts.
+ * @property {Post[]} favs - Array of favorite posts.
+ * @property {function} setFavs - Function to set the favorite posts.
+ * @property {function} addFavorite - Function to add a post to the favorites.
+ * @property {function} removeFavorite - Function to remove a post from the favorites.
+ * @property {number} totalCount - The total number of posts.
+ */
 export const HackerNewsContext = createContext({
   loadingApi: true,
   page: 0,
   setPage: (page: number) => {},
-  perPage: 10,
-  setPerPage: (perPage: number) => {},
   technologyType: "",
   setTechnologyType: (technologyType: string) => {},
   posts: [] as Post[],
@@ -22,6 +35,11 @@ export const HackerNewsContext = createContext({
   totalCount: 0,
 });
 
+/**
+ * Provides a context for the Hacker News API data and favorites management.
+ * @param children The child components to be wrapped by the provider.
+ * @returns The provider component.
+ */
 export const HackerNewsProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
@@ -62,6 +80,13 @@ export const HackerNewsProvider: React.FC<{ children: React.ReactNode }> = ({
 
   // Get from API
   useEffect(() => {
+    /**
+     * Fetches posts from the API based on the search type and technology type.
+     * Filters out posts with null values for author, story_title, story_url, and created_at.
+     * Checks if the fetched posts are already liked by the user.
+     * Sets the fetched posts and total count of pages.
+     * @returns {Promise<void>}
+     */
     const loadingPostFromApi = async () => {
       setLoadingApi(true);
 
@@ -103,8 +128,6 @@ export const HackerNewsProvider: React.FC<{ children: React.ReactNode }> = ({
         loadingApi,
         page,
         setPage,
-        perPage,
-        setPerPage,
         technologyType,
         setTechnologyType,
         posts,
