+++ src/Utils/helpers.ts
@@ -1,3 +1,9 @@
+/**
+ * Truncates a given text to a specified number of words and appends "..." at the end.
+ * @param text - The text to be truncated.
+ * @param wordLimit - The maximum number of words to be displayed.
+ * @returns The truncated text.
+ */
 export const truncateText = (text: string | null, wordLimit: number) => {
   if (text === null) {
     return "";
